# Optimized polygon move-ip. Added more tests and bencmarks.